server/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
idna==3.7
python-dotenv==1.0.1
requests==2.31.0
setuptools==68.2.2
sqlparse==0.5.0
urllib3==2.2.1
wheel==0.41.2
